refactor: policy cache
