+++ pkg/policycache/cache_test.go
@@ -2,50 +2,15 @@ package policycache
 
 import (
 	"encoding/json"
-	"fmt"
 	"testing"
 
 	kyverno "github.com/kyverno/kyverno/api/kyverno/v1"
 	"github.com/kyverno/kyverno/pkg/autogen"
-	lv1 "github.com/kyverno/kyverno/pkg/client/listers/kyverno/v1"
 	"gotest.tools/assert"
-	"k8s.io/apimachinery/pkg/labels"
 )
 
-type dummyLister struct {
-}
-
-func (dl dummyLister) List(selector labels.Selector) (ret []*kyverno.ClusterPolicy, err error) {
-	return nil, fmt.Errorf("not implemented")
-}
-
-func (dl dummyLister) Get(name string) (*kyverno.ClusterPolicy, error) {
-	return nil, fmt.Errorf("not implemented")
-}
-
-func (dl dummyLister) ListResources(selector labels.Selector) (ret []*kyverno.ClusterPolicy, err error) {
-	return nil, fmt.Errorf("not implemented")
-}
-
-// type dymmyNsNamespace struct {}
-
-type dummyNsLister struct {
-}
-
-func (dl dummyNsLister) Policies(name string) lv1.PolicyNamespaceLister {
-	return dummyNsLister{}
-}
-
-func (dl dummyNsLister) List(selector labels.Selector) (ret []*kyverno.Policy, err error) {
-	return nil, fmt.Errorf("not implemented")
-}
-
-func (dl dummyNsLister) Get(name string) (*kyverno.Policy, error) {
-	return nil, fmt.Errorf("not implemented")
-}
-
 func Test_All(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newPolicy(t)
 	//add
 	pCache.add(policy)
@@ -77,7 +42,7 @@ func Test_All(t *testing.T) {
 }
 
 func Test_Add_Duplicate_Policy(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newPolicy(t)
 	pCache.add(policy)
 	pCache.add(policy)
@@ -103,7 +68,7 @@ func Test_Add_Duplicate_Policy(t *testing.T) {
 }
 
 func Test_Add_Validate_Audit(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newPolicy(t)
 	pCache.add(policy)
 	pCache.add(policy)
@@ -128,7 +93,7 @@ func Test_Add_Validate_Audit(t *testing.T) {
 }
 
 func Test_Add_Remove(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newPolicy(t)
 	kind := "Pod"
 	pCache.add(policy)
@@ -156,7 +121,7 @@ func Test_Add_Remove(t *testing.T) {
 }
 
 func Test_Add_Remove_Any(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newAnyPolicy(t)
 	kind := "Pod"
 	pCache.add(policy)
@@ -184,7 +149,7 @@ func Test_Add_Remove_Any(t *testing.T) {
 }
 
 func Test_Remove_From_Empty_Cache(t *testing.T) {
-	pCache := newPolicyCache(nil, nil)
+	pCache := newPolicyCache()
 	policy := newPolicy(t)
 
 	pCache.remove(policy)
@@ -925,7 +890,7 @@ func newValidateEnforcePolicy(t *testing.T) *kyverno.ClusterPolicy {
 }
 
 func Test_Ns_All(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newNsPolicy(t)
 	//add
 	pCache.add(policy)
@@ -957,7 +922,7 @@ func Test_Ns_All(t *testing.T) {
 }
 
 func Test_Ns_Add_Duplicate_Policy(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newNsPolicy(t)
 	pCache.add(policy)
 	pCache.add(policy)
@@ -984,7 +949,7 @@ func Test_Ns_Add_Duplicate_Policy(t *testing.T) {
 }
 
 func Test_Ns_Add_Validate_Audit(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newNsPolicy(t)
 	pCache.add(policy)
 	pCache.add(policy)
@@ -1009,7 +974,7 @@ func Test_Ns_Add_Validate_Audit(t *testing.T) {
 }
 
 func Test_Ns_Add_Remove(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newNsPolicy(t)
 	nspace := policy.GetNamespace()
 	kind := "Pod"
@@ -1027,7 +992,7 @@ func Test_Ns_Add_Remove(t *testing.T) {
 }
 
 func Test_GVk_Cache(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newGVKPolicy(t)
 	//add
 	pCache.add(policy)
@@ -1043,7 +1008,7 @@ func Test_GVk_Cache(t *testing.T) {
 }
 
 func Test_GVK_Add_Remove(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newGVKPolicy(t)
 	kind := "ClusterRole"
 	pCache.add(policy)
@@ -1060,7 +1025,7 @@ func Test_GVK_Add_Remove(t *testing.T) {
 }
 
 func Test_Add_Validate_Enforce(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newUserTestPolicy(t)
 	nspace := policy.GetNamespace()
 	//add
@@ -1076,7 +1041,7 @@ func Test_Add_Validate_Enforce(t *testing.T) {
 }
 
 func Test_Ns_Add_Remove_User(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newUserTestPolicy(t)
 	nspace := policy.GetNamespace()
 	kind := "Deployment"
@@ -1094,7 +1059,7 @@ func Test_Ns_Add_Remove_User(t *testing.T) {
 }
 
 func Test_Mutate_Policy(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newMutatePolicy(t)
 	//add
 	pCache.add(policy)
@@ -1113,7 +1078,7 @@ func Test_Mutate_Policy(t *testing.T) {
 }
 
 func Test_Generate_Policy(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newgenratePolicy(t)
 	//add
 	pCache.add(policy)
@@ -1130,7 +1095,7 @@ func Test_Generate_Policy(t *testing.T) {
 }
 
 func Test_NsMutate_Policy(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy := newMutatePolicy(t)
 	nspolicy := newNsMutatePolicy(t)
 	//add
@@ -1155,7 +1120,7 @@ func Test_NsMutate_Policy(t *testing.T) {
 }
 
 func Test_Validate_Enforce_Policy(t *testing.T) {
-	pCache := newPolicyCache(dummyLister{}, dummyNsLister{})
+	pCache := newPolicyCache()
 	policy1 := newValidateAuditPolicy(t)
 	policy2 := newValidateEnforcePolicy(t)
 	pCache.add(policy1)
